add accelerated networking + gen2 as first class capabilities
also move capabilities to a separate file, and update data tests for new capabilities.
